+++ R/plot_metrics.R
@@ -161,19 +161,25 @@ draw_radar_plot <- function(score_frame, type) {
     }
 
     data  <- rbind(f1_row, recall_row, precision_row, accuracy_row, auc_row)
-    data  <- data.frame(names(score_frame[-c(1)]), data)
+    data  <- data.frame(names(score_frame[-c(1, 2, 3, 4)]), data)
     title <- 'Model comparison'
+
+    colnames(data) <- c('Metric', score_frame$name)
 
   } else {
     r2_row <- NULL
+    print(names(score_frame))
+    print(score_frame)
     if ('r2' %in% names(score_frame)) {
       r2_row <- as.data.frame(t(as.numeric(unlist(score_frame['r2']))))
     }
     data  <- r2_row
     title <- 'R2 comparison'
-  }
 
-  colnames(data) <- score_frame$engine
+    colnames(data) <- score_frame$name
+  }
+  print(data)
+  #colnames(data) <- c('Metric', score_frame$name)
 
   ggradar::ggradar(plot.data                = data,
                    base.size                = 10,