+++ tests/testthat/test-check_data.R
@@ -80,18 +80,17 @@ test_that('test-check_data', {
   expect_output(check_cor(df_iris2, y_iris2), cor_iris2)
   expect_output(check_cor(df_test, y_test), cor_test)
 
-  no_outliers <- '<U\\+2714> No outliers in the dataset.'
-  out_lisbon  <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 145 146 196 44 5 51 57 58 59 60 61 62 63 64 69 75 76 77 78 ;'
-  out_compas  <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 102 108 1181 1209 1321 1401 1403 1406 1408 1417 1422 1443 1468 1526 1532 1561 157 1596 1630 1681 173 1814 1820 1830 1865 1920 1924 1950 2080 2099 210 2105 2168 2264 2301 2336 2348 2410 2417 2423 2444 2453 2503 2504 2526 2544 2611 2648 2680 273 2744 2792 2829 2858 2871 2872 2873 2888 2979 3043 3050 3104 3107 3138 3204 3207 322 3229 3250 326 3280 3314 3333 3360 3394 34 3534 356 3594 3620 3714 3762 3803 3830 3872 3873 3923 393 4076 4083 4085 4086 4091 4104 4111 4172 424 425 4274 4303 4378 4394 4484 4492 455 4709 4720 4740 4882 4931 4962 4973 5067 5133 5164 5193 5231 5254 5261 5283 5286 5332 5350 5351 539 5411 5497 5512 555 556 5705 5830 588 5934 5959 5979 6003 6005 6023 603 6080 6161 630 642 674 707 709 739 787 846 897 904 932 ;'
-  out_iris2   <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 16 ;'
-  out_adult   <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 1000 102 107 11 113 127 132 144 149 158 161 171 173 183 190 193 198 201 204 209 216 222 223 225 231 24 245 249 273 286 29 300 313 322 328 33 364 38 382 388 405 406 409 414 415 417 427 431 435 464 469 487 490 509 511 521 53 535 537 574 589 590 598 620 633 638 652 666 669 670 698 702 705 722 730 733 743 75 78 782 787 800 802 825 828 85 873 874 9 915 916 919 930 933 936 938 94 948 964 97 973 991 ;'
-  out_test    <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 160 209 365 369 395 434 481 491 559 6 791 795 796 804 82 ;'
+  no_outliers  <- '<U\\+2714> No outliers in the dataset.'
+  out_lisbon   <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 145 146 196 44 5 51 57 58 59 60 61 62 63 64 69 75 76 77 78 ;'
+  out_over_50  <- '<U\\+2716> There are more than 50 possible outliers in the data set, so we are not printing them. They are returned in the output as a vector.'
+  out_iris2    <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 16 ;'
+  out_test     <- '<U\\+2716> These obserwation migth be outliers due to their numerical columns values: \n 160 209 365 369 395 434 481 491 559 6 791 795 796 804 82 ;'
 
   expect_output(check_outliers(df_iris), no_outliers)
   expect_output(check_outliers(df_lisbon), out_lisbon)
-  expect_output(check_outliers(df_compas), out_compas)
+  expect_output(check_outliers(df_compas), out_over_50)
   expect_output(check_outliers(df_iris2), out_iris2)
-  expect_output(check_outliers(df_adult), out_adult)
+  expect_output(check_outliers(df_adult), out_over_50)
   expect_output(check_outliers(df_test), out_test)
 
   balanced       <- '<U\\+2714> Dataset is balanced.'